Serialization module refactor
