+++ include/boost/mysql/impl/internal/protocol/impl/bit_deserialization.hpp
@@ -5,13 +5,13 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BOOST_MYSQL_IMPL_INTERNAL_PROTOCOL_BIT_DESERIALIZATION_HPP
-#define BOOST_MYSQL_IMPL_INTERNAL_PROTOCOL_BIT_DESERIALIZATION_HPP
+#ifndef BOOST_MYSQL_IMPL_INTERNAL_PROTOCOL_IMPL_BIT_DESERIALIZATION_HPP
+#define BOOST_MYSQL_IMPL_INTERNAL_PROTOCOL_IMPL_BIT_DESERIALIZATION_HPP
 
 #include <boost/mysql/field_view.hpp>
 #include <boost/mysql/string_view.hpp>
 
-#include <boost/mysql/impl/internal/protocol/serialization.hpp>
+#include <boost/mysql/impl/internal/protocol/impl/deserialization_context.hpp>
 
 #include <boost/endian/conversion.hpp>
 
@@ -26,7 +26,7 @@ namespace detail {
 // string_lenenc layer, this function is in charge of just parsing the binary payload. The length of
 // the BIT value depends on how the type was defined in the table (e.g. BIT(14) will send a 2 byte
 // value; BIT(54) will send a 7 byte one). Values are sent as big-endian.
-inline deserialize_errc deserialize_bit(string_view from, field_view& to) noexcept
+inline deserialize_errc deserialize_bit(string_view from, field_view& to)
 {
     std::size_t num_bytes = from.size();
     if (num_bytes < 1 || num_bytes > 8)
