+++ include/boost/mysql/impl/internal/protocol/impl/span_string.hpp
@@ -5,8 +5,8 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BOOST_MYSQL_IMPL_INTERNAL_PROTOCOL_SPAN_STRING_HPP
-#define BOOST_MYSQL_IMPL_INTERNAL_PROTOCOL_SPAN_STRING_HPP
+#ifndef BOOST_MYSQL_IMPL_INTERNAL_PROTOCOL_IMPL_SPAN_STRING_HPP
+#define BOOST_MYSQL_IMPL_INTERNAL_PROTOCOL_IMPL_SPAN_STRING_HPP
 
 #include <boost/mysql/string_view.hpp>
