+++ include/boost/mysql/impl/internal/sansio/message_reader.hpp
@@ -12,7 +12,7 @@
 #include <boost/mysql/error_code.hpp>
 
 #include <boost/mysql/impl/internal/protocol/constants.hpp>
-#include <boost/mysql/impl/internal/protocol/protocol.hpp>
+#include <boost/mysql/impl/internal/protocol/deserialization.hpp>
 #include <boost/mysql/impl/internal/sansio/read_buffer.hpp>
 
 #include <boost/asio/coroutine.hpp>
