+++ include/boost/mysql/impl/internal/sansio/reset_connection.hpp
@@ -14,7 +14,8 @@
 
 #include <boost/mysql/detail/algo_params.hpp>
 
-#include <boost/mysql/impl/internal/protocol/protocol.hpp>
+#include <boost/mysql/impl/internal/protocol/deserialization.hpp>
+#include <boost/mysql/impl/internal/protocol/serialization.hpp>
 #include <boost/mysql/impl/internal/sansio/connection_state_data.hpp>
 #include <boost/mysql/impl/internal/sansio/sansio_algorithm.hpp>
 #include <boost/mysql/impl/internal/sansio/set_character_set.hpp>
