+++ test/unit/include/test_unit/create_execution_processor.hpp
@@ -18,8 +18,6 @@
 #include <boost/mysql/detail/execution_processor/execution_processor.hpp>
 #include <boost/mysql/detail/resultset_encoding.hpp>
 
-#include <boost/mysql/impl/internal/protocol/protocol.hpp>
-
 #include <cstddef>
 
 #include "test_unit/create_meta.hpp"
