+++ test/unit/include/test_unit/printing.hpp
@@ -15,7 +15,7 @@
 #include <boost/mysql/detail/rows_iterator.hpp>
 
 #include <boost/mysql/impl/internal/protocol/capabilities.hpp>
-#include <boost/mysql/impl/internal/protocol/db_flavor.hpp>
+#include <boost/mysql/impl/internal/protocol/constants.hpp>
 #include <boost/mysql/impl/internal/sansio/next_action.hpp>
 
 #include <boost/test/unit_test.hpp>
