+++ test/unit/pch.hpp
@@ -14,12 +14,14 @@
 #include <boost/asio/post.hpp>
 #include <boost/asio/ssl/stream.hpp>
 #include <boost/asio/ssl/stream_base.hpp>
+#include <boost/assert.hpp>
 #include <boost/config.hpp>
 #include <boost/core/detail/string_view.hpp>
 #include <boost/core/span.hpp>
 #include <boost/describe/class.hpp>
 #include <boost/describe/members.hpp>
 #include <boost/describe/operators.hpp>
+#include <boost/endian/conversion.hpp>
 #include <boost/mp11.hpp>
 #include <boost/system/error_category.hpp>
 #include <boost/system/error_code.hpp>
