+++ test/unit/test/execution_processor/execution_processor_helpers.hpp
@@ -13,8 +13,6 @@
 
 #include <boost/mysql/detail/coldef_view.hpp>
 
-#include <boost/mysql/impl/internal/protocol/protocol.hpp>
-
 #include <boost/test/unit_test.hpp>
 
 #include "test_common/check_meta.hpp"
